+++ src/skland/gui/main-window.cpp
@@ -41,8 +41,8 @@ MainWindow::MainWindow(int width, int height, const char *title, int flags)
     output_size = output->current_mode_size();  // The current screen size
   }
 
-  int total_width = std::max(this->width(), output_size.width);
-  int total_height = std::max(this->height(), output_size.height);
+  int total_width = std::max((int) (this->width()), output_size.width);
+  int total_height = std::max((int) (this->height()), output_size.height);
   if (!IsFrameless()) {
     total_width += surface()->margin().lr();
     total_height += surface()->margin().tb();
